requirements.txt
streamlit>=1.42.0
pandas
google-cloud-bigquery
db-dtypes
pandas-gbq
qrcode[pil]
Pillow
pyzbar
altair
bcrypt
extra-streamlit-components
